m1l1 first app and test, ignore external files
