+++ .gitignore
@@ -1,0 +1,6 @@
+/.idea/
+/.gradle/
+build/
+!gradle/wrapper/gradle-wrapper.jar
+!**/src/main/**/build/
+!**/src/test/**/build/
