--- .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Thu Jan 11 10:46:27 MSK 2024
-gradle.version=7.6.1
