--- .idea/yatool.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="YaTool.Application.Settings">
-    <option name="yaToolPath" value="$ARCADIA_ROOT$/ya" />
-    <option name="autoDownloadYaTool" value="true" />
-  </component>
-</project>
