+++ README.md
@@ -1,0 +1,3 @@
+# Marketplace
+Marketplace is the world's largest online store offering a huge range of products, from electronics and home appliances to clothing, shoes, accessories and food.
+Founded in 2024, Marketplace has become a global e-commerce platform, focusing on customer convenience, large selection and fast delivery of orders.
