+++ gradle.properties
@@ -1,1 +1,2 @@
 kotlin.code.style=official
+kotlinVersion=1.9.21
