+++ m1l1-first/src/main/kotlin/Main.kt
@@ -1,0 +1,6 @@
+package ru.otus.otuskotlin.marketplace.m1l1
+
+fun main(args: Array<String>) {
+    println("Hello World!")
+    println("Program arguments: ${args.joinToString()}")
+}
