+++ m1l1-first/src/main/kotlin/Summator.kt
@@ -1,0 +1,5 @@
+package ru.otus.otuskotlin.marketplace.m1l1
+
+class Summator {
+    fun sum(a: Int, b: Int): Int = a + b
+}
