+++ settings.gradle.kts
@@ -1,9 +1,13 @@
 pluginManagement {
-    repositories {
-        mavenCentral()
-        gradlePluginPortal()
+    plugins {
+        val kotlinVersion: String by settings
+        kotlin("jvm") version kotlinVersion
     }
-    
 }
-rootProject.name = "Marketplace"
 
+plugins {
+    id("org.gradle.toolchains.foojay-resolver-convention") version "0.5.0"
+}
+rootProject.name = "otuskotlin-marketplace-202312"
+
+include("m1l1-first")
